Manage changes in keywords in newer urllib3 versions
The module urllib3 has changed some keywords, apparently when going from 1.25 to 1.26
Since genx does not require a specific version of this module (it is inherited from requests),
this breaking change raises errors on users with newer versions of this library. This fix uses
the approach sugested in urllib3's issue 2057.